chore: rename coordinateScaler file
